Moved metatable to own files. As they will be getting very large. I.E. mtVector2.cpp
