+++ lua.h
@@ -1,6 +1,28 @@
 #ifndef LUA_H
 #define LUA_H
 
+// These 2 macros helps us in constructing general metamethods.
+// We can use "lua" as a "Lua" pointer and arg1, arg2, ..., arg5 as Variants objects
+// Check examples in createVector2Metatable
+#define LUA_LAMBDA_TEMPLATE(_f_) \
+ [](lua_State* inner_state) -> int {                      \
+     lua_pushstring(inner_state,"__Lua");                 \
+     lua_rawget(inner_state,LUA_REGISTRYINDEX);           \
+     Lua* lua = (Lua*) lua_touserdata(inner_state,-1);;   \
+     lua_pop(inner_state,1);                              \
+     Variant arg1 = lua->getVariant(1);                             \
+     Variant arg2 = lua->getVariant(2);                             \
+     Variant arg3 = lua->getVariant(3);                             \
+     Variant arg4 = lua->getVariant(4);                             \
+     Variant arg5 = lua->getVariant(5);                             \
+     _f_                                                            \
+}
+ 
+#define LUA_METAMETHOD_TEMPLATE( lua_state , metatable_index , metamethod_name , _f_ )\
+lua_pushstring(lua_state,metamethod_name); \
+lua_pushcfunction(lua_state,LUA_LAMBDA_TEMPLATE( _f_ )); \
+lua_settable(lua_state,metatable_index-2);
+
 #include "core/object/ref_counted.h"
 #include "core/core_bind.h"
